Extract gamestate and code that controls it
