+++ main.js
@@ -4,49 +4,26 @@ let stampit = require('stampit');
 let THREE = require('three');
 let WebGLRenderer = require('./src/webgl_renderer.js');
 let Player1 = require('./src/player1.js');
-let sounds = require('./src/sounds.js');
 let Selector = require('./src/selector.js');
 let MouseState = require('./src/mouse_state.js');
 let MouseStream = require('./src/mouse_stream.js');
 let Debug = require('./src/debug.js');
 let {Goomba} = require('./src/enemies.js')
 let gameRules = require('./src/game_rules.js').create();
+let gameState = require('./src/game_state.js');
+let gameStateController = require('./src/game_state_controller.js');
 let levelData = require('./src/level_data.js');
 
-let {inputStream, menuStream, jumpStream, inputState} = require("./src/input_stream.js");
-
-let gameState = {
-  paused: false
-}
-
-menuStream.onValue((x) => {
-  sounds.pauseSound.pause();
-  sounds.pauseSound.currentTime = 0;
-  sounds.pauseSound.play();
-  document.getElementById("game-wrapper").classList.toggle("paused");
-  gameState.paused = ! gameState.paused;
-});
-
 let renderer = WebGLRenderer.create({canvas: document.getElementById('game-canvas')});
-renderer.loadLevel(levelData['1-1']);
-
 let player1  = Player1.create({renderer: renderer, position: new THREE.Vector2(5, 4) });
-let selector = Selector.create({renderer: renderer, position: new THREE.Vector2(0, 0)});
-
-let mouseState = MouseState.create({canvasId: "game-canvas"});
-
-mouseState.addListener(function(x,y) {
-  x = Math.floor(x/32 - renderer.camera.position.x);
-  y = Math.floor(y/32);
-  selector.moveTo(x,y);
-});
-
 let entities = []
 
 for(var i=0; i<1; i++) {
   let goom = Goomba.create({renderer: renderer, position: new THREE.Vector2(20+i*2, 6)})
   entities.push(goom);
 }
+
+renderer.loadLevel(levelData['1-1']);
 
 function render() {
   let dt = 1/60;
@@ -60,28 +37,25 @@ function render() {
       renderer.camera.position.x = (4 - player1.position.x);
     }
 
-    selector.update(dt);
     player1.update(dt);
     renderer.render(dt);
 
     gameRules.update(player1, entities, gameState);
   }
+
   Debug('mario-x', player1.position.x);
   Debug('mario-y', player1.position.y);
 
   requestAnimationFrame(render);
 }
 
-window.player = player1;
-
 render();
 
+let selector = Selector.create({renderer: renderer, position: new THREE.Vector2(0, 0)});
+let mouseState = MouseState.create({canvasId: "game-canvas"});
 
-document.getElementById('play-music').onclick = function() {
-  sounds.groundTheme.play();
-};
-
-document.getElementById('stop-music').onclick = function() {
-  sounds.groundTheme.pause();
-  sounds.groundTheme.currentTime = 0;
-};
+mouseState.addListener(function(x,y) {
+  x = Math.floor(x/32 - renderer.camera.position.x);
+  y = Math.floor(y/32);
+  selector.moveTo(x,y);
+});
