+++ src/game_state.js
@@ -1,0 +1,27 @@
+let sounds = require('./sounds.js');
+
+let gameState = {
+  paused: false,
+  musicPlaying: false,
+
+  playMusic: function() {
+    musicPlaying = true;
+    sounds.groundTheme.play();
+  },
+
+  stopMusic: function() {
+    musicPlaying = false;
+    sounds.groundTheme.pause();
+    sounds.groundTheme.currentTime = 0;
+  },
+
+  togglePause: function() {
+    sounds.pauseSound.pause();
+    sounds.pauseSound.currentTime = 0;
+    sounds.pauseSound.play();
+    document.getElementById("game-wrapper").classList.toggle("paused");
+    this.paused = ! this.paused;
+  }
+};
+
+module.exports = gameState;
